upbit-grid/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
python-dotenv==1.0.1
pydantic==2.10.5
sqlmodel==0.0.22
httpx==0.28.1
pyjwt==2.10.1
passlib[bcrypt]==1.7.4
# Upbit client
pyupbit==0.2.34
